pyproject.toml
[tool.poetry]
name = "mindtools"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2"
boto3 = "^1.34.106"
botocore = ">=1.12.26"
cairocffi = ">=0.9.0"
cairosvg = ">=1.0.22"
certifi = ">=2018.10.15"
cffi = ">=1.11.5"
chardet = ">=3.0.4"
contextlib2 = ">=0.5.5"
cssselect2 = ">=0.2.1"
decorator = ">=4.3.0"
diff-match-patch = ">=20121119"
django-appconf = ">=1.0.2"
django-compressor = ">=2.2"
django-constance = ">=2.3.1"
django-defender = ">=0.7.0"
django-extensions = ">=2.1.3"
django-filer = "3.1.1"
django-import-export = ">=1.1.0"
django-mptt = ">=0.8.7"
django-polymorphic = "3.1"
django-request = ">=1.5.4"
django-reversion = ">=3.0.0"
django-settings-export = ">=1.2.1"
django-sitetree = ">=1.12.0"
django-storages = ">=1.6.5"
django-user-agents = ">=0.3.2"
django-xforwardedfor-middleware = ">=1.1"
djangorestframework = ">=3.6.4"
docutils = ">=0.14"
easy-thumbnails = ">=2.3"
et-xmlfile = ">=1.0.1"
gevent = ">=1.3.7"
greenlet = ">=0.4.15"
gunicorn = ">=19.9.0"
html5lib = ">=1.0.1"
httplib2 = ">=0.11.3"
huey = ">=1.2.1"
idna = ">=2.7"
jdcal = ">=1.4"
jmespath = ">=0.9.3"
jsonfield = ">=3.1.0,<3.2.0"
lxml = ">=4.2.5"
mistune = ">=0.8.4"
odfpy = ">=1.3.6"
openpyxl = ">=2.5.8"
pdfrw = ">=0.4"
pillow = ">=5.3.0"
plivo = ">=4.1.3"
psycopg2-binary = "2.9.9"
pycparser = ">=2.19"
pyparsing = ">=2.2.2"
pyphen = ">=0.9.5"
python-dateutil = ">=2.7.3,<2.8.1"
pyyaml = ">=3.13"
raven = ">=6.9.0"
rcssmin = ">=1.0.6"
redis = ">=2.10.6,<=3.5"
requests = ">=2.19.1"
requests-file = ">=1.4.3"
rjsmin = ">=1.0.12"
rules = ">=2.0"
s3transfer = ">=0.1.13"
six = ">=1.11.0"
tablib = ">=0.12.1"
tinycss2 = ">=0.6.1"
tldextract = ">=2.2.0"
twilio = ">=5.7.0"
typing = ">=3.6.6"
ua-parser = ">=0.8.0"
unicodecsv = ">=0.14.1"
unidecode = ">=0.4.21,<1.1"
urllib3 = ">=1.23, <2"
user-agents = ">=1.1.0"
virtualenv = ">=16.0.0"
weasyprint = ">=0.42.3"
xlrd = ">=1.1.0"
xlwt = ">=1.3.0"
django-npm = ">=1.0.0"
future = ">=0.18.2"
django-admin-sortable = "^2.3"
django-ratelimit = "^4.1.0"
backports-csv = "^1.0.7"
django-ipware = "^7.0.1"
django-configurations = {extras = ["database"], version = "^2.5.1"}
psycopg2 = "^2.9.9"
sentry-sdk = {extras = ["django"], version = "^2.5.1"}
html2text = "^2024.2.26"
django-author = "^1.2.0"
ijson = "^3.3.0"

[tool.poetry.dependencies.django-suit-v2-pm]
git = "https://github.com/pulse-mind/django-suit.git"
branch = "v2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
